idaho/backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
idna==2.7
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.20.0
urllib3==1.24.1
